Fold VarHandle.handleTable and its element
handleTable is an immutable array with immutable content, thus we can
fold loads of it and its elements if the containing VarHandle object is
immutable. We will set isArrayWithConstantElements bit for handleTable
when creating an entry in KnownObjectTable. Then VP sees the bit and
will improve array shadows from handleTable to ImmutableArrayShadows,
which is to be detected by transformIndirectLoadChain when trying to
fold an array shadow.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>