Improve fieldwatch implementation across all codegens
This commit implements general cleanup in the rdbar and wrtbar
evaluators across the x86 and Z codegens. This includes refactoring
common code between the x86 and Z codegens into the common codegenerator
to improve synergy, consolidating routines where it makes sense to do so,
and improving the general design of the fieldwatch feature from the codegen
perspective.

There are documentation and bug fixes implemented as well that improve
the robustness and maintainability of the fieldwatch feature.

Finally, new snippet types are implemented at the common codegen level that
will specifically be used for fieldwatch data.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>